+++ src/utils.js
@@ -3,21 +3,25 @@ const fs = require('fs');
 const path = require('path');
 const chalk = require('chalk');
 
-function getConfig() {
-  const configPath = path.join(process.cwd(), '.deploy');
+function getConfig(filename = '.deploy') {
   try {
+    const configPath = path.join(process.cwd(), filename);
     if (fs.existsSync(configPath)) {
       return JSON.parse(fs.readFileSync(configPath, 'utf-8'));
     }
     return {};
   } catch (error) {
-    return {};
+    console.error(chalk.red(error.message));
+    process.exit(1);
   }
 }
 
+const privateKeyPath = path.join(process.cwd(), 'private.pem');
+const publicKeyPath = path.join(process.cwd(), 'public.pem');
+
 function encrypt(str) {
   try {
-    const rsa = new NodeRSA(fs.readFileSync('public.pem', 'utf8'), { encryptionScheme: 'pkcs1' });
+    const rsa = new NodeRSA(fs.readFileSync(publicKeyPath, 'utf8'), { encryptionScheme: 'pkcs1' });
     return rsa.encrypt(str, 'base64');
   } catch (error) {
     console.error(chalk.red(error.message));
@@ -27,7 +31,7 @@ function encrypt(str) {
 
 function decrypt(str) {
   try {
-    const rsa = new NodeRSA(fs.readFileSync('private.pem', 'utf8'), { encryptionScheme: 'pkcs1' });
+    const rsa = new NodeRSA(fs.readFileSync(privateKeyPath, 'utf8'), { encryptionScheme: 'pkcs1' });
     return rsa.decrypt(str, 'utf8');
   } catch (error) {
     console.error(chalk.red(error.message));