+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0(2017-06-15)
+### Features
+* add token `rsa` verify
+* **client** add `--config` Specify the configuration file, default `.deploy`
+
 # 0.1.1(2017-06-09)
 ### Bug Fixes
 * **lunux** `: No such file or directory`